requirements.txt
streamlit>=1.30.0
fastapi>=0.110.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
langchain>=0.1.0
langchain-community>=0.0.20
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
transformers>=4.36.0
torch>=2.0.0
ollama>=0.1.6
